Refactor toward feature-based architecture for backup and search.
Extract backup import helpers and message search service into dedicated modules, and move persona picker UI into the personas feature.
