+++ lib/features/conversations/data/models/message_search_hit.dart
@@ -1,0 +1,17 @@
+import '../../../../core/models/enums.dart';
+
+class MessageSearchHit {
+  const MessageSearchHit({
+    required this.messageId,
+    required this.conversationId,
+    required this.conversationTitle,
+    required this.snippet,
+    required this.role,
+  });
+
+  final String messageId;
+  final String conversationId;
+  final String conversationTitle;
+  final String snippet;
+  final MessageRole role;
+}
